Refactor move confirmation
